some fixes

+++ testing/ProductionCode2/ProductionCode2/FileService.cs
@@ -13,41 +13,78 @@ namespace ProductionCode
         readonly string _filterExtension = ".tmp";
         readonly string _outputFilename = "/backup.tmp";
         readonly string _errorLogFilename = "/error.log";
-        readonly string _removeFilesListFilename = "/error.log";
+        readonly string _removeFilesListFilename = "/ToRemove.txt";
 
         public int MergeTemporaryFiles(string dir)
         {
-            IEnumerable<string> files = Directory.GetFiles(dir).Where(t => t.EndsWith(_filterExtension));
+            if (!this.ValidationDirectory(dir))
+                throw new DirectoryNotFoundException();
+
+            IEnumerable<string> files = this.GetFiterExtensionFiles(Directory.GetFiles(dir));
 
             if (files.Count() == 0)
                 return 0;
 
-            File.WriteAllText(Path.Combine(dir, _outputFilename), files.Aggregate((t,k) => string.Concat(t,k)));
+            File.WriteAllText(Path.Combine(dir, _outputFilename), this.GenerateOutputFileText(files.Select(t => File.ReadAllText(t))));
             files.ToList().ForEach(t => File.Delete(t));
 
-            return files.Count();
+            return this.GetCountValidFiles(files);
         }
+
+        protected virtual bool ValidationDirectory(string dir)
+            => Directory.Exists(dir);
+
+        private int GetCountValidFiles(IEnumerable<string> files)
+            => files.Count();
+
+        private string GenerateOutputFileText(IEnumerable<string> fileTexts)
+            => fileTexts.Aggregate((t, k) => string.Concat(t, k));
+
+        private IEnumerable<string> GetFiterExtensionFiles(IEnumerable<string> files)
+            => files.Where(t => t.EndsWith(_filterExtension));
 
         public int RemoveTemporaryFiles(string dir)
         {
-            IEnumerable<string> files = File.ReadAllLines(Path.Combine(dir, _removeFilesListFilename));
+            if (!this.ValidationDirectory(dir))
+                throw new DirectoryNotFoundException();
+
+            IEnumerable<string> files = this.GetFiles(dir);
 
             long sizeDeletedFiles = default(int);
             files.ToList()
-                .ForEach(t => 
+                .ForEach(t =>
                 {
                     try
                     {
-                        sizeDeletedFiles =+ new FileInfo(t).Length;
-                        File.Delete(t);
+                        if (!this.ValidationFile(t))
+                            throw new FileNotFoundException();
+
+                        sizeDeletedFiles += this.GetSizeInBytes(t);
                     }
                     catch
                     {
-                        File.AppendAllText(Path.Combine(dir, _errorLogFilename), $"Фaйл {Path.GetFileNameWithoutExtension(t)} не был найден.{Environment.NewLine}");
+                        this.WriteErrorMessageToLog(dir, t);
                     }
                 });
 
             return (int)sizeDeletedFiles;
+        }
+
+        protected virtual bool ValidationFile(string filePath)
+            => File.Exists(filePath);
+
+        protected virtual IEnumerable<string> GetFiles(string dir)
+            => File.ReadAllLines(Path.Combine(dir, _removeFilesListFilename));
+
+        protected virtual void WriteErrorMessageToLog(string dir, string t)
+            => File.AppendAllText(Path.Combine(dir, _errorLogFilename), $"Фaйл {Path.GetFileNameWithoutExtension(t)} не был найден.{Environment.NewLine}");
+
+        protected virtual long GetSizeInBytes(string t)
+        {
+            long sizeDeletedFiles = new FileInfo(t).Length;
+            File.Delete(t);
+
+            return sizeDeletedFiles;
         }
     }
 }
